+++ skorch/tests/test_regressor.py
@@ -4,6 +4,8 @@ Only contains tests that are specific for regressor subclasses.
 
 """
 
+from functools import partial
+
 import numpy as np
 import pytest
 from sklearn.base import clone
@@ -20,6 +22,12 @@ class TestNeuralNetRegressor:
     def module_cls(self):
         from skorch.toy import make_regressor
         return make_regressor(dropout=0.5)
+
+    @pytest.fixture(scope='module')
+    def module_pred_1d_cls(self):
+        from skorch.toy import MLPModule
+        # Module that returns 1d predictions
+        return partial(MLPModule, output_units=1, squeeze_output=True)
 
     @pytest.fixture(scope='module')
     def net_cls(self):
@@ -57,9 +65,9 @@ class TestNeuralNetRegressor:
     def test_clone(self, net_fit):
         clone(net_fit)
 
-    def test_fit(self, net_fit):
-        # fitting does not raise anything
-        pass
+    def test_fit(self, net_fit, recwarn):
+        # fitting does not raise anything and does not warn
+        assert not recwarn.list
 
     @pytest.mark.parametrize('method', INFERENCE_METHODS)
     def test_not_fitted_raises(self, net_cls, module_cls, data, method):
@@ -91,17 +99,6 @@ class TestNeuralNetRegressor:
         for row in net_fit.history:
             assert expected_keys.issubset(row)
 
-    def test_target_1d_raises(self, net, data):
-        X, y = data
-        with pytest.raises(ValueError) as exc:
-            net.fit(X, y.flatten())
-        assert exc.value.args[0] == (
-            "The target data shouldn't be 1-dimensional but instead have "
-            "2 dimensions, with the second dimension having the same size "
-            "as the number of regression targets (usually 1). Please "
-            "reshape your target data to be 2-dimensional "
-            "(e.g. y = y.reshape(-1, 1).")
-
     def test_predict_predict_proba(self, net_fit, data):
         X = data[0]
         y_pred = net_fit.predict(X)
@@ -123,3 +120,57 @@ class TestNeuralNetRegressor:
         multioutput_net.fit(X, y)
         r2_score = multioutput_net.score(X, y)
         assert r2_score <= 1.
+
+    def test_dimension_mismatch_warning(self, net_cls, module_cls, data, recwarn):
+        # When the target and the prediction have different dimensionality, mse
+        # loss will broadcast them, calculating all pairwise errors instead of
+        # only sample-wise. Since the errors are averaged at the end, there is
+        # still a valid loss, which makes the error hard to spot. Thankfully,
+        # torch gives a warning in that case. We test that this warning exists,
+        # otherwise, skorch users could run into very hard to debug issues
+        # during training.
+        net = net_cls(module_cls)
+        X, y = data
+        X, y = X[:100], y[:100].flatten()  # make y 1d
+        net.fit(X, y)
+
+        w0, w1 = recwarn.list  # one warning for train, one for valid 
+        # The warning comes from PyTorch, so checking the exact wording is prone to
+        # error in future PyTorch versions. We thus check a substring of the
+        # whole message and cross our fingers that it's not changed.
+        msg_substr = (
+            "This will likely lead to incorrect results due to broadcasting. "
+            "Please ensure they have the same size"
+        )
+        assert msg_substr in str(w0.message)
+        assert msg_substr in str(w1.message)
+
+    def test_fitting_with_1d_target_and_pred(
+            self, net_cls, module_cls, data, module_pred_1d_cls, recwarn
+    ):
+        # This test relates to the previous one. In general, users should fit
+        # with target and prediction being 2d, even if the 2nd dimension is just
+        # 1. However, in some circumstances (like when using BaggingRegressor,
+        # see next test), having the ability to fit with 1d is required. In that
+        # case, the module output also needs to be 1d for correctness.
+        X, y = data
+        X, y = X[:100], y[:100]  # less data to run faster
+        y = y.flatten()
+
+        net = net_cls(module_pred_1d_cls)
+        net.fit(X, y)
+        assert not recwarn.list
+
+    def test_bagging_regressor(
+            self, net_cls, module_cls, data, module_pred_1d_cls, recwarn
+    ):
+        # https://github.com/skorch-dev/skorch/issues/972
+        from sklearn.ensemble import BaggingRegressor
+
+        net = net_cls(module_pred_1d_cls)  # module output should be 1d too
+        X, y = data
+        X, y = X[:100], y[:100]  # less data to run faster
+        y = y.flatten()  # make y 1d or else sklearn will complain
+        regr = BaggingRegressor(net, n_estimators=2, random_state=0)
+        regr.fit(X, y)  # does not raise
+        assert not recwarn.list  # ensure there is no broadcast warning from torch